requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2
django-mathfilters==1.0.0
django-multiselectfield==0.1.13
django-multiupload==0.6.1
djangorestframework==3.16.0
idna==3.10
pillow==11.2.1
requests==2.32.3
sqlparse==0.5.3
urllib3==2.4.0
whitenoise==6.9.0
